[feature] - code review comments: code partition, text/keyboard blockage
