+++ App.js
@@ -1,108 +1,8 @@
 import React from 'react';
-import {SafeAreaView, StyleSheet, View, Platform} from 'react-native';
-import Home from "./Components/Home";
-import {StackNavigator, TabNavigator} from "react-navigation";
-import {purple, white} from "./utils/colors";
-import {FontAwesome, Ionicons} from '@expo/vector-icons';
-import AddEditQuestion from "./Components/AddEditQuestion";
-import DeckView from "./Components/DeckView";
-import AddDeck from "./Components/AddDeck";
-import QuizContainer from "./Components/QuizContainer";
-import Notifications from "./Components/Notifications";
+import {SafeAreaView, StyleSheet, View} from 'react-native';
+import {white} from "./utils/colors";
 import {setLocalNotification} from "./utils/helpers";
-
-const Tabs = TabNavigator({
-		Home: {
-			screen: Home,
-			navigationOptions: {
-				tabBarLabel: 'Home',
-				tabBarIcon: ({tintColor}) => <Ionicons name='ios-home'
-																							 color={tintColor}
-																							 size={30}/>
-			}
-		},
-		AddDeck: {
-			screen: AddDeck,
-			navigationOptions: {
-				tabBarLabel: 'Add Deck',
-				tabBarIcon: ({tintColor}) => <FontAwesome name='plus-square'
-																									color={tintColor}
-																									size={30}/>
-			}
-		},
-		Notifications: {
-			screen: Notifications,
-			navigationOptions: {
-				tabBarLabel: 'Notifications',
-				tabBarIcon: ({tintColor}) => <Ionicons name='ios-notifications'
-																							 color={tintColor}
-																							 size={30}/>
-			}
-		},
-	}, {
-		navigationOptions: {
-			headers: null
-		}
-	},
-	{
-		tabBarOptions: {
-			activeTintColor: Platform.OS === 'ios' ? purple : white,
-			style: {
-				height: 56,
-				backgroundColor: Platform.OS === 'ios' ? white : purple,
-				shadowColor: 'rgba(0,0,0,0.24)',
-				shadowOffset: {
-					width: 0,
-					height: 3
-				},
-				shadowRadius: 6,
-				shadowOpacity: 1,
-			}
-		}
-	}
-);
-
-const MainNavigator = StackNavigator({
-	Home: {
-		screen: Tabs,
-	},
-	AddDeck: {
-		screen: AddDeck,
-		navigationOptions: {
-			headerTintColor: white,
-			headerStyle: {
-				backgroundColor: purple
-			}
-		}
-	},
-	DeckView: {
-		screen: DeckView,
-		navigationOptions: {
-			headerTintColor: white,
-			headerStyle: {
-				backgroundColor: purple
-			}
-		}
-	},
-	QuizContainer: {
-		screen: QuizContainer,
-		navigationOptions: {
-			headerTintColor: white,
-			headerStyle: {
-				backgroundColor: purple
-			}
-		}
-	},
-	AddEditQuestion: {
-		screen: AddEditQuestion,
-		navigationOptions: {
-			headerTintColor: white,
-			headerStyle: {
-				backgroundColor: purple
-			}
-		}
-	},
-});
+import MainNavigator from "./Components/MainNavigator";
 
 export default class App extends React.Component {
 	componentDidMount(){
