+++ Components/QuizCard.js
@@ -78,8 +78,8 @@ const styles = StyleSheet.create({
 		justifyContent: "center",
 	},
 	flipCard: {
-		width: 400,
-		height: 400,
+		width: 200,
+		height: 200,
 		alignItems: 'center',
 		justifyContent: 'center',
 		backgroundColor: orange,
